+++ package.json
@@ -14,7 +14,6 @@
     "axios": "^1.1.3",
     "lodash": "^4.17.21",
     "react": "^18.2.0",
-    "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-intersection-observer": "^9.4.1",
