+++ src/apis/axiosInstance.ts
@@ -4,10 +4,14 @@ export const authAxiosInstance = axios.create({
   baseURL: process.env.REACT_APP_AUTH_BASE_URL,
   headers: {
     'Content-type': 'application/json',
-    accecpt: 'application/json',
+    Accecpt: 'application/json',
   },
 });
 
 export const boardAxiosInstance = axios.create({
   baseURL: process.env.REACT_APP_BOARD_BASE_URL,
+});
+
+export const authAxios = axios.create({
+  baseURL: process.env.REACT_APP_AUTH_BASE_URL,
 });
